+++ geopolrisk-py/assessment/operations.py
@@ -1,17 +1,3 @@
-# Copyright 2020-2021 by Anish Koyamparambath and University of Bordeaux. All Rights Reserved.
-# Permission to use, copy, modify, and distribute this software and its
-# documentation for any purpose and without fee is hereby granted,
-# provided that the above copyright notice appear in all copies and that
-# both that copyright notice and this permission notice appear in
-# supporting documentation, and that the name of Anish Koyamparambath (AK) or
-# University of Bordeaux (UBx) will not be used in advertising or publicity pertaining
-# to distribution of the software without specific, written prior permission.
-# BOTH AK AND UBx DISCLAIMS ALL WARRANTIES WITH REGARD TO THIS SOFTWARE, INCLUDING
-# ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL
-# BOTH AK AND UBx BE LIABLE FOR ANY SPECIAL, INDIRECT OR CONSEQUENTIAL DAMAGES OR
-# ANY DAMAGES WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER
-# IN AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT
-# OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
 
 from .__init__ import (
@@ -26,14 +12,7 @@ from .__init__ import (
 )
 
 _outputfile = exportfile
-from .core import (
-    regions,
-    COMTRADE_API,
-    InputTrade,
-    WTA_calculation,
-    productionQTY,
-    GeoPolRisk,
-)
+from .core import *
 from .Exceptions.warningsgprs import *
 import itertools, sqlite3, pandas as pd, time
 
@@ -49,235 +28,7 @@ _columns = [
     "Weighted Trade AVerage",
 ]
 outputList = []
-# Extract list of all data
-HS = _price.HS.to_list()
-Resource = _price.Resource.to_list()
-Country = _reporter.Country.to_list()
-ISO = _reporter.ISO.to_list()
 
-
-# The function converts resource names and country names to the HS commodity codes
-# and country ISO codes.
-
-def convertCodes(resource, country, *args, **kwargs):
-    
-    # Verify direction
-    def check_variables(A, B):
-        if (isinstance(A, list) and 
-            isinstance(B, list) and 
-            all(isinstance(element, str) 
-                for element in A) and 
-                all(isinstance(element, str) for element in B)):
-            return "Numeric"
-        elif (isinstance(A, int) and isinstance(B, int) or 
-              (isinstance(A, list) and isinstance(B, list) and 
-               all(isinstance(element, int) for element in A) and 
-               all(isinstance(element, int) for element in B))):
-            return "Text"
-        else:
-            return None
-    
-    def return_variables(X, A, B):
-        if isinstance(X, list):
-            X_transform = []
-            for i, n in enumerate(X):
-                try:
-                    idx = A.index(n)
-                    X_transform.append(B[idx])
-                except Exception as e:
-                    logging.debug("Failed to transform")
-                    logging.debug(e)
-                    logging.debug("Transformation failed for " + str(n))
-            return X_transform
-        else:
-            try:
-                idx = A.index(X)
-                X_transform = B[idx]
-            except Exception as e:
-                    logging.debug("Failed to transform")
-                    logging.debug(e)
-                    logging.debug("Transformation failed for " + str(X))
-            return X_transform
-
-    direction = check_variables(resource, country)    
-    if direction == "Numeric":
-        ResourceCX = return_variables(resource, Resource, HS)
-        CountryCX = return_variables(country, Country, ISO)
-    elif direction == "Text":
-        ResourceCX = return_variables(resource, HS, Resource)
-        CountryCX = return_variables(country, ISO, Country)
-    else:
-        logging.debug("Failed to verify direction")
-        logging.debug(direction)
-    return ResourceCX, CountryCX
-
-# Verify if the calculation is already stored in the database to avoid recalculation
-def sqlverify(*args):
-    resource, country, year, recyclingrate, scenario = (
-        args[0],
-        args[1],
-        args[2],
-        args[3],
-        args[4],
-    )
-    sqlstatement = (
-        "SELECT geopolrisk, hhi, wta, geopol_cf FROM recordData WHERE country = '"
-        + country
-        + "' AND resource= '"
-        + resource
-        + "' AND year = '"
-        + str(year)
-        + "' AND recycling_rate = '"
-        + str(recyclingrate)
-        + "' AND scenario = '"
-        + str(scenario)
-        + "';"
-    )
-    try:
-        row = execute_query(sqlstatement)
-    except InputError:
-        logging.debug(sqlstatement)
-    if not row:
-        return None
-    else:
-        outputList.append(
-            [
-                str(year),
-                resource,
-                country,
-                recyclingrate,
-                scenario,
-                row[0][0],
-                row[0][3],
-                row[0][1],
-                row[0][2],
-            ]
-        )
-        return True
-
-
-# Store the calculation into a database
-# Please note that database is not included in the library, this is automatically created
-# in the documents directory of the operating system
-def recorddata(*args):
-    (
-        resource,
-        country,
-        year,
-        recyclingrate,
-        scenario,
-        GPRS,
-        CF,
-        HHI,
-        WTA,
-        HSCODE,
-        ISO,
-        Log,
-    ) = (
-        args[0],
-        args[1],
-        args[2],
-        args[3],
-        args[4],
-        args[5],
-        args[6],
-        args[7],
-        args[8],
-        args[9],
-        args[10],
-        args[11],
-    )
-    sqlstatement = (
-        "INSERT INTO recordData (country, resource, year, recycling_rate,"
-        " scenario, geopolrisk, hhi, wta, geopol_cf, resource_hscode, iso, log_ref) VALUES ('"
-        "" + country + "','" + resource + "','" + str(year) + ""
-        "','" + str(recyclingrate) + "','" + str(scenario) + "','"
-        "" + str(GPRS) + "','" + str(HHI) + "','" + str(WTA) + ""
-        "','" + str(CF) + "','" + str(HSCODE) + "','" + str(ISO) + "',"
-        "'" + str(Log) + "');"
-    )
-    try:
-        row = execute_query(sqlstatement)
-        return True
-    except Exception as e:
-        logging.debug(e)
-
-
-# Updates the database using new values.
-# This function doesnt override the calculaiton
-def updatedata(*args):
-    # Verify
-    resource, country, year, recyclingrate, scenario = (
-        args[0],
-        args[1],
-        args[2],
-        args[3],
-        args[4],
-    )
-    GPRS, CF, HHI, WTA, Log = args[5], args[6], args[7], args[8], args[9]
-    sqlstatement = (
-        "SELECT geopolrisk, hhi, wta, geopol_cf from "
-        "recordData WHERE country = '"
-        + country
-        + "' AND resource= '"
-        + resource
-        + "' AND year = '"
-        + str(year)
-        + "' AND recycling_rate = '"
-        + str(recyclingrate)
-        + "' AND scenario = '"
-        + str(scenario)
-        + "';"
-    )
-    row = execute_query(sqlstatement)
-    dataframe = pd.DataFrame(row, columns=["geopolrisk", "hhi", "wta", "geopol_cf"])
-    if dataframe.shape[0] > 2:
-        logging.debug("Multiple data records found!")
-        raise DataRecordError
-        return None
-    elif dataframe.shape[0] == 0:
-        logging.debug("No data records found!")
-        raise DataRecordError
-        return None
-    else:
-        if float(dataframe.iloc[0]["wta"]) == float(WTA):
-            logging.debug("NO change in trade data detected! No SQL executed.")
-            return None
-        else:
-            sqlstatement = (
-                "UPDATE recordData SET hhi= '"
-                + str(HHI)
-                + "', wta ='"
-                + str(WTA)
-                + "', geopolrisk='"
-                + str(GPRS)
-                + "', geopol_cf= '"
-                + str(CF)
-                + "', log_ref='"
-                + str(Log)
-                + "' WHERE country = '"
-                + country
-                + "' AND resource= '"
-                + resource
-                + "' AND year = '"
-                + str(year)
-                + "' AND recycling_rate = '"
-                + str(recyclingrate)
-                + "' AND scenario = '"
-                + str(scenario)
-                + "';"
-            )
-            norow = execute_query(sqlstatement)
-            logging.debug("Database update sucessfully!")
-
-
-# Function to calculate the geopolitical supply risk and CFs for the midpoint indicator
-# A single point calculation function. It takes in the list of resources (HS code commodities),
-# list of countries (ISO 3 digit code) and list of years. The recycling rate and scenario are provided as
-# float and integers.
-# The argument 'record' for database overrides the capacity to update the database. That means if
-# a calculation exists, a repeated calculation is not done. 'update' as an argument updates the database
-# instead of recording.
 def gprs_comtrade(
     resourcelist, countrylist, 
     yearlist, recyclingrate, 
@@ -300,7 +51,6 @@ def gprs_comtrade(
             resource, country = convertCodes(I[0], I[1], 2)
             verify = sqlverify(resource, country, I[2], recyclingrate, scenario)
         except Exception as e:
-            print(e)
             logging.debug(e)
             raise IncompleteProcessFlow
